pyproject.toml
[tool.poetry]
name = "fyp"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
absl-py = "^2.1.0"
pandas = "1.5.3"
dacite = "^1.8.1"
nbformat = "^5.10.4"
dill = "^0.3.8"
sacrebleu = "^2.4.2"
astor = "^0.8.1"
folium = "^0.16.0"
seaborn = "^0.13.2"
bokeh = "^3.4.1"
plotly = "^5.21.0"
matplotlib = "3.7.1"
chart-studio = "^1.1.0"
seqio = "^0.0.19"
diff-match-patch = "^20230430"
ipython-genutils = "^0.2.0"
cloudpickle = "^3.0.0"
statsmodels = "^0.14.2"
openpyxl = "^3.1.2"
kaggle = "^1.6.12"
vega = "^4.0.0"
openai = "^1.23.2"
python-dotenv = "^1.0.1"
transformers = "^4.40.1"
accelerate = "^0.29.3"
codebleu = "^0.6.0"
tiktoken = "^0.6.0"
blobfile = "^2.1.1"
tenacity = "^8.3.0"
dvc = "^3.50.2"
dvc-s3 = "^3.2.0"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
tabulate = "^0.9.0"
networkx = "^3.3"
torchmetrics = "^1.4.0.post0"
evaluate = "^0.4.2"
nltk = "^3.8.1"
rouge-score = "^0.1.2"
tree-sitter-python = "^0.21.0"
faiss-cpu = "^1.8.0"
llama-index-vector-stores-faiss = "^0.1.2"
gradio = "^4.36.1"
instructor = "^1.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
